Moves files and fix imports errors
