+++ W40K/Company.py
@@ -1,6 +1,7 @@
 from MyPack.Utilities import *
-from W40K.Class.Weapons import *
-from W40K.Class.FuncAndTables.Functions import *
+#from W40K.Class.Weapons import *
+from W40K.FuncAndTables.Functions import *
+from W40K.FuncAndTables.TableValues import *
 
 class Company:
     def __init__(self,Unit=None,Equipement=[]):
@@ -22,9 +23,6 @@ class Company:
         self.Breakthrought = float()
         self.Defense = float()
         self.HOI4_Profil()
-    def check_lists(self):
-        for el in self.Equipement:
-            assert type(el) is Weapon , "Each element of Equipement list must be a :Weapon class:"
     def HOI4_Profil(self):
         if self.Unit is None:
             pass
